pyproject.toml
[project]
name = "aide"
version = "0.1.0"
description = "AI detection service"
readme = "README.md"
requires-python = ">=3.10,<3.12"
dependencies = [
    "datasets>=4.2.0",
    "fastapi>=0.119.0",
    "httpx>=0.28.1",
    "pandas>=2.3.3",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.2",
    "tokenizers>=0.22.1",
    "transformers>=4.57.1",
    "uvicorn>=0.37.0",
]
